LayoutTests/imported/w3c/web-platform-tests/fetch/api/abort/general-serviceworker.https-expected.txt

CONSOLE MESSAGE: line 14: Unhandled Promise Rejection: UnknownError: serviceWorker.getRegistration() is not yet implemented

FAIL General fetch abort tests in a service worker serviceWorker.getRegistration() is not yet implemented
